refactor: label/tooltip/warning ⇒ `TakeNowDropdown`
Extract more content into `TakeNowModal` component by creating
`TakeNowDropdown`, a Take Now-specific wrapper for `ParticipantDropdown`
